pyproject.toml
[project]
name = "h3blog"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "alipay-sdk-python>=3.7.826",
    "bs4>=0.0.2",
    "dotenv>=0.9.9",
    "flask>=3.1.2",
    "flask-caching>=2.3.1",
    "flask-login>=0.6.3",
    "flask-migrate>=4.1.0",
    "flask-sitemap>=0.4.0",
    "flask-sqlalchemy>=3.1.1",
    "flask-wtf>=1.2.2",
    "geoip2>=5.1.0",
    "email_validator",
    "html2text>=2025.4.15",
    "markdown>=3.9",
    "markdown-checklist>=0.4.4",
    "openpyxl>=3.1.5",
    "oss2>=2.19.1",
    "pillow>=12.0.0",
    "pygments>=2.19.2",
    "pymysql>=1.1.2",
    "redis>=7.0.0",
    "wechatpayv3>=2.0.1",
]
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

# 新增：显式指定只打包 app 包
[tool.setuptools.packages.find]
include = ["app*"]  # 包含 app 及其子包（如 app.models, app.views）
exclude = ["logs", "uploads", "tests", ".*"]  # 排除非代码目录

# 测试依赖组
[project.optional-dependencies]
test = [
    "pytest",         # 核心测试框架
    "pytest-html",    # HTML 报告（固定版本避免样式乱掉）
    "allure-pytest", # Allure 插件（版本要和 CLI 匹配）
    "requests"       # 你的测试里用到的请求库
]